Feature/rest duration between exercise (#47)
* Add restDuration on ExerciseGroupDataModel

* Display restDuration on exerciseGroup

* Add restDuration on exerciseGroup create

* Convert param type in front end not in back

---------

+++ src/Domain/DTO/DataModel/Workout/ExerciseDataModel.php
@@ -22,32 +22,32 @@ class ExerciseDataModel implements DataModelInterface
     #[ORM\Column(type: Types::INTEGER, nullable: true)]
     public ?int $targetReps = null;
 
-    #[ORM\Column(type: Types::INTEGER, nullable: true)]
-    public ?int $targetWeight = null;
+    #[ORM\Column(type: Types::FLOAT, nullable: true)]
+    public ?float $targetWeight = null; // in Kg
 
     #[ORM\Column(type: Types::INTEGER, nullable: true)]
-    public ?int $targetDuration = null;
+    public ?int $targetDuration = null; // in seconds
 
-    #[ORM\Column(type: Types::INTEGER, nullable: true)]
-    public ?int $targetDistance = null;
+    #[ORM\Column(type: Types::FLOAT, nullable: true)]
+    public ?float $targetDistance = null; // in Km
 
-    #[ORM\Column(type: Types::INTEGER, nullable: true)]
-    public ?int $targetSpeed = null;
+    #[ORM\Column(type: Types::FLOAT, nullable: true)]
+    public ?float $targetSpeed = null; // in Km/h
 
     #[ORM\Column(type: Types::INTEGER, nullable: true)]
     public ?int $reps = null;
 
-    #[ORM\Column(type: Types::INTEGER, nullable: true)]
-    public ?int $weight = null;
+    #[ORM\Column(type: Types::FLOAT, nullable: true)]
+    public ?int $weight = null; // in Kg
 
     #[ORM\Column(type: Types::INTEGER, nullable: true)]
-    public ?int $duration = null;
+    public ?int $duration = null; // in seconds
 
-    #[ORM\Column(type: Types::INTEGER, nullable: true)]
-    public ?int $distance = null;
+    #[ORM\Column(type: Types::FLOAT, nullable: true)]
+    public ?int $distance = null; // in Km
 
-    #[ORM\Column(type: Types::INTEGER, nullable: true)]
-    public ?int $speed = null;
+    #[ORM\Column(type: Types::FLOAT, nullable: true)]
+    public ?int $speed = null; // in Km/h
 
     #[ORM\Column(type: Types::BOOLEAN)]
     public bool $isCompleted = false;
